requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==3.2.16
django-bootstrap-v5==1.0.4
django-ckeditor==6.0.0
django-js-asset==2.0.0
Pillow==8.4.0
pytz==2022.5
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
